requirements.txt
numpy
matplotlib
astropy
scipy
wquantiles
colorio
pyspectacle
